Reverting MATAR-Kokkoification of elements/
Focus MATAR-Kokkoification in elements-kokkos/
